Refactor: reorganización de espacios de nombres
Se reorganizaron las rutas de los espacios de nombres para las
interfaces `IEmpresaRepository` e `IEmpresaService`, moviéndolas
a `PruebaTecnica.Data.Repositories.Impl` y
`PruebaTecnica.Services.Impl`, respectivamente.

Se actualizaron las referencias en los archivos afectados:
- `EmpresaRepository.cs`, `EmpresaService.cs`, `FormCompanyEditor.cs`,
  `FormCompanyList.cs` y `Program.cs`.

Se eliminaron las definiciones previas de las interfaces en
`PruebaTecnica.Interfaces` y se añadieron nuevas definiciones en
los espacios de nombres correspondientes.
