+++ Data/Repositories/Impl/IEmpresaRepository.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PruebaTecnica.Interfaces
+namespace PruebaTecnica.Data.Repositories.Impl
 {
     public interface IEmpresaRepository
     {
