+++ Forms/FormCompanyEditor.cs
@@ -1,7 +1,6 @@
 ﻿using PruebaTecnica.Data;
-using PruebaTecnica.Interfaces;
 using PruebaTecnica.Models;
-using PruebaTecnica.Services;
+using PruebaTecnica.Services.Impl;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
